create TCBase wrapper to add common functionality of TC wrappers
