+++ metplus/wrappers/tcrmw_wrapper.py
@@ -15,7 +15,7 @@ import os
 from ..util import ti_calculate, ti_get_hours_from_relativedelta
 from ..util import do_string_sub, get_lead_sequence
 from ..util import parse_var_list, sub_var_list
-from . import RuntimeFreqWrapper
+from . import TCBaseWrapper
 
 '''!@namespace TCRMWWrapper
 @brief Wraps the TC-RMW tool
@@ -23,7 +23,7 @@ from . import RuntimeFreqWrapper
 '''
 
 
-class TCRMWWrapper(RuntimeFreqWrapper):
+class TCRMWWrapper(TCBaseWrapper):
     RUNTIME_FREQ_DEFAULT = 'RUN_ONCE_PER_INIT_OR_VALID'
     RUNTIME_FREQ_SUPPORTED = ['RUN_ONCE_PER_INIT_OR_VALID']
 
@@ -140,13 +140,7 @@ class TCRMWWrapper(RuntimeFreqWrapper):
                                              'TC_RMW_VALID_HOUR',
                                              ])
 
-        self.add_met_config(name='compute_tangential_and_radial_winds', data_type='bool')
-        self.add_met_config(name='u_wind_field_name', data_type='string')
-        self.add_met_config(name='v_wind_field_name', data_type='string')
-        self.add_met_config(name='tangential_velocity_field_name', data_type='string')
-        self.add_met_config(name='tangential_velocity_long_field_name', data_type='string')
-        self.add_met_config(name='radial_velocity_field_name', data_type='string')
-        self.add_met_config(name='radial_velocity_long_field_name', data_type='string')
+        self.add_met_config_tc_wind()
 
         c_dict['VAR_LIST_TEMP'] = parse_var_list(self.config,
                                                  data_type='FCST',
@@ -249,50 +243,6 @@ class TCRMWWrapper(RuntimeFreqWrapper):
         self._set_lead_list(time_info, lead_seq)
 
         return time_info
-
-    def _set_data_field(self, time_info):
-        """!Get list of fields from config to process. Build list of field info
-            that are formatted to be read by the MET config file. Set DATA_FIELD
-            item of c_dict with the formatted list of fields.
-            Args:
-                @param time_info time dictionary to use for string substitution
-                @returns True if field list could be built, False if not.
-        """
-        field_list = sub_var_list(self.c_dict['VAR_LIST_TEMP'], time_info)
-        if not field_list:
-            self.log_error("Could not get field information from config.")
-            return False
-
-        all_fields = []
-        for field in field_list:
-            field_list = self.get_field_info(d_type='FCST',
-                                             v_name=field['fcst_name'],
-                                             v_level=field['fcst_level'],
-                                             )
-            if field_list is None:
-                self.log_error(f'Could not get field info from {field}')
-                return False
-
-            all_fields.extend(field_list)
-
-        data_field = ','.join(all_fields)
-        self.env_var_dict['METPLUS_DATA_FIELD'] = f'field = [{data_field}];'
-        return True
-
-    def _set_lead_list(self, time_info, lead_seq):
-        # set LEAD_LIST to list of forecast leads used
-        if lead_seq == [0]:
-            return
-
-        lead_list = []
-        for lead in lead_seq:
-            lead_hours = (
-                ti_get_hours_from_relativedelta(lead,
-                                                valid_time=time_info['valid'])
-            )
-            lead_list.append(f'"{str(lead_hours).zfill(2)}"')
-
-        self.env_var_dict['METPLUS_LEAD_LIST'] = f"lead = [{', '.join(lead_list)}];"
 
     def set_command_line_arguments(self, time_info):
         if self.c_dict['CONFIG_FILE']:
